refactor: reorganize emoji exports into separate file
